requirements.txt
mysql-connector-python==8.0.23
frozendict==1.2
